The feauture to add a new todo-item to the todo list was implemented.
The business rules which guarantee a todo-item is valid was added.

The endpoints to control the requests which have to do with the todo list
are ready to use.

repository pattern was implemented in the task-management module, currently
the unique operation the repository do is to add a new todo-item.