Bug in the verification of the title was removed.
The repository function to verify the existence of a title
in a todo item was taking account  only the id of the user but
not the id of the list. For that reason if the title was existing
in one of the to do items of any of the user's lists, the client
was getting an error in which the server advised a title duplication.

Also, some debug logs were removed from the code.